% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grm.bin.input.R
\name{grm.bin.input}
\alias{grm.bin.input}
\title{grm import function}
\usage{
grm.bin.input(file, size = 4)
}
\arguments{
\item{file}{name of the binary grm file. No default.}

\item{size}{byte size used (typically gcta uses 4). See eponymous parameter of function \code{\link{readBin}}.}
}
\value{
object of type matrix

\code{grmsem.bin.input} imports a binary GCTA GRM (.bin) file and transforms it into a symmetric matrix
}
\description{
This function imports genetic relationship matrices (GRMs)
in binary format as e.g. stored by the GCTA software command \code{gcta64 --grm test --make-grm-bin --out test}
}
