% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grm.input.R
\name{grm.input}
\alias{grm.input}
\title{grm import function}
\usage{
grm.input(file)
}
\arguments{
\item{file}{the name of the gz compressed grm file. No default.}
}
\value{
\code{grmsem.input} imports a zipped GCTA GRM (.gz) file and transforms it into a symmetric matrix
}
\description{
This function imports genetic relationship matrices (GRMs) in gz format as e.g.
stored by the GCTA software command \code{gcta64 --grm test --make-grm-gz --out test}
}
\keyword{grmsem}
